Consolidate portfolio.py helpers into aux_functions.py
portfolio.py collapses to:
  - run_portfolio orchestrator
  - lazy module __getattr__ + __dir__ re-exporting helpers from aux_functions

All ten helpers — add_ecdf, _build_industry_daily_returns,
_build_industry_monthly_returns, _build_hml_lms, portfolios, regional_data,
_build_regional_loop, _stack_outputs, _write_filtered, _write_split_by_key —
now live in aux_functions.py alongside the rest of the pipeline's building
blocks (mirrors the main.py / aux_functions.py split on the characteristics
side).

Importing `jkp.data.portfolio` no longer pulls duckdb/ibis at module-import
time; those load only on first re-export access or on calling run_portfolio
(which has a function-local import of the helpers it dispatches to).

The Polars Sortedness UserWarning suppression is scoped inside add_ecdf via
warnings.catch_warnings() rather than mutating the global filter list at
aux_functions import time.
